Reorganise forcefield data classes and lib.
