+++ src/classes/pairpotential.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "base/parameters.h"
-#include "data/ff.h"
+#include "data/ff/ff.h"
 #include "math/data1d.h"
 #include "math/interpolator.h"
 #include "templates/list.h"
