+++ src/data/ff/angleterm.cpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2021 Team Dissolve and contributors
 
-#include "data/ffangleterm.h"
-#include "data/ff.h"
-#include "data/ffatomtype.h"
+#include "data/ff/angleterm.h"
+#include "data/ff/ff.h"
+#include "data/ff/atomtype.h"
 
 ForcefieldAngleTerm::ForcefieldAngleTerm(std::string_view typeI, std::string_view typeJ, std::string_view typeK,
                                          SpeciesAngle::AngleFunction form, const std::vector<double> parameters)
