+++ src/data/ff/bondterm.cpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2021 Team Dissolve and contributors
 
-#include "data/ffbondterm.h"
-#include "data/ff.h"
-#include "data/ffatomtype.h"
+#include "data/ff/bondterm.h"
+#include "data/ff/ff.h"
+#include "data/ff/atomtype.h"
 
 ForcefieldBondTerm::ForcefieldBondTerm(std::string_view typeI, std::string_view typeJ, SpeciesBond::BondFunction form,
                                        const std::vector<double> parameters)
