+++ src/data/ff/ff.cpp
@@ -1,19 +1,19 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2021 Team Dissolve and contributors
 
-#include "data/ff.h"
+#include "data/ff/ff.h"
 #include "classes/atomtype.h"
 #include "classes/box.h"
 #include "classes/coredata.h"
 #include "classes/species.h"
 #include "classes/speciesatom.h"
 #include "classes/speciesbond.h"
-#include "data/ffangleterm.h"
-#include "data/ffatomtype.h"
-#include "data/ffbondterm.h"
-#include "data/ffimproperterm.h"
-#include "data/ffparameters.h"
-#include "data/fftorsionterm.h"
+#include "data/ff/angleterm.h"
+#include "data/ff/atomtype.h"
+#include "data/ff/bondterm.h"
+#include "data/ff/improperterm.h"
+#include "data/ff/parameters.h"
+#include "data/ff/torsionterm.h"
 
 /*
  * Set Up
