+++ src/data/ff/ff.h
@@ -8,12 +8,12 @@
 #include "classes/speciesimproper.h"
 #include "classes/speciestorsion.h"
 #include "data/elements.h"
-#include "data/ffangleterm.h"
-#include "data/ffatomtype.h"
-#include "data/ffbondterm.h"
-#include "data/ffimproperterm.h"
-#include "data/ffparameters.h"
-#include "data/fftorsionterm.h"
+#include "data/ff/angleterm.h"
+#include "data/ff/atomtype.h"
+#include "data/ff/bondterm.h"
+#include "data/ff/improperterm.h"
+#include "data/ff/parameters.h"
+#include "data/ff/torsionterm.h"
 #include "templates/optionalref.h"
 #include <algorithm>
 #include <functional>
