+++ src/data/ff/library.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "data/ff.h"
+#include "data/ff/ff.h"
 #include <memory>
 #include <string>
