+++ src/data/ff/ludwig/ntf2.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "data/ff.h"
+#include "data/ff/ff.h"
 
 // Forward Declarations
 /* none */
