+++ src/data/ff/ludwig/py5.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "data/ff.h"
+#include "data/ff/ff.h"
 
 // Forward Declarations
 /* none */
