+++ src/data/ff/oplsaa2005/base.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "data/ff.h"
+#include "data/ff/ff.h"
 #include "templates/optionalref.h"
 
 // Forward Declarations
