+++ src/gui/selectspecieswidget.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "data/fflibrary.h"
+#include "data/ff/library.h"
 #include "gui/ui_selectspecieswidget.h"
 #include "templates/list.h"
 #include <QWidget>
